bugfix: fixed lag spike when downloading for the first time enhancement: added download progress overlay code: changed alle byte variables to int, because it doesn't make a difference

+++ src/main/java/com/vinurl/api/VinURLSound.java
@@ -11,7 +11,7 @@ import net.minecraft.world.World;
 import static com.vinurl.util.Constants.*;
 
 public class VinURLSound {
-	private static final byte JUKEBOX_RANGE = 64;
+	private static final int JUKEBOX_RANGE = 64;
 
 	public static void play(World world, ItemStack stack, BlockPos position) {
 		NbtComponent nbt = stack.get(DataComponentTypes.CUSTOM_DATA);
