+++ src/main/java/com/vinurl/client/ClientConfig.java
@@ -37,7 +37,7 @@ public class ClientConfig {
 
 	@SectionHeader("audioSettings")
 	@RangeConstraint(min = 1, max = 60)
-	public byte maxAudioInMinutes = 60;
+	public int maxAudioInMinutes = 60;
 
 	public AudioQuality audioBitrate = AudioQuality.MEDIUM;
